requirements.txt
astral==0.8.1
awsebcli==3.4.5
beautifulsoup4==4.4.0
cement==2.4.0
Django==1.9.4
django-blog-zinnia==0.15.2
django-chart-tools==1.0
django-chartit==0.2.7
django-contrib-comments==1.6.1
django-debug-toolbar==1.5
django-google-charts==0.1.1
django-mptt==0.7.4
django-qsstats-magic==0.7.2
django-tagging==0.4
django-xmlrpc==0.1.5
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
jmespath==0.7.1
Markdown==2.6.2
mots-vides==2015.5.11
MySQL-python==1.2.5
pathspec==0.3.3
Pillow==2.9.0
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
PyYAML==3.11
reportlab==3.2.0
requests==2.6.2
six==1.9.0
sqlparse==0.2.1
texttable==0.8.3
virtualenv==13.0.3
websocket-client==0.32.0
